Change projectile damage to be of int type. Enemy self-inflicts damage when hit with a projectile.

+++ Assets/Scripts/Weapons/QuadshotWeapon.cs
@@ -8,7 +8,7 @@ public class QuadshotWeapon : MonoBehaviour, IWeapon
     public float speed = 15;
     public float timeBetweenShots = 1.5f;
     public float projectileLifetime = 10;
-    public float damage = 1;
+    public int damage = 1;
     public float forwardOffset = 1.5f;
     public float upwardOffset = 1.5f;
     private float timer = 0;
